Convert typeahead to have a core that can be used by other components (e.g. address lookup)
