+++ components/02-form-elements/address/_address-input.scss
@@ -1,0 +1,15 @@
+.address-input {
+  &__search {
+    display: none;
+  }
+
+  &--search & {
+    &__manual {
+      display: none;
+    }
+
+    &__search {
+      display: block;
+    }
+  }
+}
