+++ components/02-form-elements/address/address.hbs
@@ -1,32 +1,56 @@
-<fieldset class="field">
+<fieldset class="field address-input js-address">
   <legend class="field__legend mars">What is your address?</legend>
-  <div class="field__item field__item--full-width">
-    <label class="label venus" for="address-line-1">Building and street <span class="u-vh">line 1 of 2</span></label>
-    <input id="address-line-1" class="input input--text input--width-20" name="address-line-1" type="text" />
+
+  <div class="address-input__manual">
+    <div class="field__item field__item--full-width">
+      <label class="label venus" for="address-line-1">Building and street <span class="u-vh">line 1 of 2</span></label>
+      <input id="address-line-1" class="input input--text input--width-20" name="address-line-1" type="text" />
+    </div>
+
+    <div class="field__item field__item--full-width">
+      <label class="u-vh" for="address-line-2">Building and street line 2 of 2</label>
+      <input id="address-line-2" class="input input--text input--width-20" name="address-line-2" type="text" />
+    </div>
+
+    <div class="field__item">
+      <label class="label venus" for="address-town">Town or City</label>
+      <input id="address-town" class="input input--text" name="address-town" type="text" />
+    </div>
+
+    <div class="field__item">
+      <label class="label venus" for="address-county">County</label>
+      <input id="address-county" class="input input--text" name="address-county" type="text" />
+    </div>
+
+    <div class="field__item">
+      <label class="label venus" for="address-postcode">Postcode</label>
+      <input id="address-postcode" class="input input--text input--width-3" name="address-postcode" type="text" />
+    </div>
+
+    <div class="field__item">
+      Or
+      <button class="btn btn--secondary btn--small js-address-search-btn" type="button">Search for an address</button>
+    </div>
   </div>
 
-  <div class="field__item field__item--full-width">
-    <label class="u-vh" for="address-line-2">Building and street line 2 of 2</label>
-    <input id="address-line-2" class="input input--text input--width-20" name="address-line-2" type="text" />
+  <div class="address-input__search">
+    <div class="field__item">
+      <div class="typeahead js-address-typeahead" data-content="{{json typeahead.content}}">
+        <label class="label venus" for="address-typeahead">Please type in your address below</label>
+        <div class="typeahead__combobox js-typeahead-combobox" role="combobox" aria-haspopup="listbox" aria-expanded="false">
+          <input class="typeahead__preview js-typeahead-preview input input--text" readonly="true" aria-hidden="true" tabindex="-1" />
+          <input id="address-typeahead" class="typeahead__input js-typeahead-input input input--text" name="address" type="text" autocomplete="false" aria-autocomplete="list" aria-controls="{{typeahead.listbox_id}}" aria-describedby="{{typeahead.instructions_id}}" />
+          <div class="u-vh" id="{{typeahead_instructions_id}}">{{typeahead.instructions}}</div>
+          <ul class="typeahead__listbox js-typeahead-listbox" role="listbox" id="{{typeahead.listbox_id}}"></ul>
+          <div class="u-vh js-typeahead-aria-status" aria-live="polite" role="status"></div>
+        </div>
+      </div>
+    </div>
+
+    <div class="field__item">
+      Or
+      <button class="btn btn--secondary btn--small js-address-manual-btn" type="button">Manually enter address</button>
+    </div>
   </div>
 
-  <div class="field__item">
-    <label class="label venus" for="address-town">Town or City</label>
-    <input id="address-town" class="input input--text" name="address-town" type="text" />
-  </div>
-
-  <div class="field__item">
-    <label class="label venus" for="address-county">County</label>
-    <input id="address-county" class="input input--text" name="address-county" type="text" />
-  </div>
-
-  <div class="field__item">
-    <label class="label venus" for="address-postcode">Postcode</label>
-    <input id="address-postcode" class="input input--text input--width-3" name="address-postcode" type="text" />
-  </div>
-
-  <div class="field__item">
-    Or
-    <button class="btn btn--secondary btn--small" type="button">Search for address</button>
-  </div>
 </fieldset>
